BallOptions and PaddleOptions stored integer positions. Now they don't.
This is very much a bug. These positions should be double floating point
to be consistent with the rest of the engine.

+++ src/server/Object.h
@@ -29,11 +29,11 @@ namespace pong
   };
   struct PaddleOptions
   {
-    math::vector<int> destination;
+    math::vector<double> destination;
   };
   struct BallOptions
   {
-    math::vector<int> velocity;
+    math::vector<double> velocity;
   };
 
   /*!
